feat: initial spike
This commit adds the initial work done towards creating a Patternfly
based theme for Backstage.  The theme bridges the Patternfly color
palette and some of the spacing from the @patternfly/react-tokens
project to Material theming and applies a number of component and
typography overrides.  This initial commit adds the most basic vite
based build with the most minimum dependencies to build.