feat: switch to monorepo and add storybook
This change converts this repo into a monorepo with the theme in one
package and adds a new package that's set up to get it's stories out of
the upstream Backstage repository, assuming BACKSTAGE_SOURCE_ROOT is set
properly in the environment.  The storybook is set up to provide the
patternfly theme for each story, however it appears some more
investigation needs to be done for the theme to fully load properly
within this new storybook instance.