fix: ensure theme palette is loaded
This change updates the build configurations for each package to ensure
that the theme doesn't bundle it's dependencies and just builds the
theme code into it's module.  This commit updates the storybook
configuration so that the backstage and material dependencies are built
into a single vendor bundle to avoid potentially differeng copies of
material libraries being loaded.