fix: remove react router
Some stories in the upstream contain code that requires the presence of
react router, however several other upstream stories also include react
router already.  This removes a stopgap put in place to handle the
stories that were affected by the first case and favors letting the
stories that are set up to rely on react router to work properly.