chore: bring up theme updates
This commit brings over the changes originally applied to the Janus
showcase app to bring this copy of the theme up to date.  The storybook
configuration is removed, this effort was unsuccessful in getting the
theme to apply with this setup.  The monorepo setup is also removed in
favor of simplicity.  Subsequent updates will focus on improving the
readme and ensuring this theme can be built and easily installed into an
existing backstage instance.